change in ModelUtilities method name from getTarget to getTargetLabel and setTarget to setTargetLabel and the local method variables from location to targetLabel
add in ModelUtilities the methods getAllNodeTemplates and getAllRelationshipTemplates

use in ModelUtilities in method getIncomingRelationshipTemplates and getOutgoingRelationshipTemplates the method getAllRelationshipTemplates instead of getNodeTemplateorRelationshipTemplate

add in Splitting the method CheckValidTopology, getSuccessorsOfNodeTemplates, computeTransitiveClosure, and computeNodeForTrnasitiveClosure

delete incomplete method split

change in Splitting the method name from split2 to split, add statement to clone the Topology, change local variable name from precedessorsLocation to precedessorsTargetLabel. Add statement to clone a newNode and to give the newNode a name and id and the same for newOutgoingRelationship and newIncomingRelationship. Add in the for loop for incomingRelationships that only incomingRelationship from nodes, which are predecessors which have the same TargetLabel, and from nodes, which are not predecessors, are attached to the newNode.